Total number or basal area of conspecific and heterospecific neighbors.

Calculates the total number or basal area of conspecific and heterospecific neighbors at a radius r with edge effect correction. Most arguments aim to describe the type of neighbors to count. NeighborDensities() counts neighbors for each of the individuals in censdata.

NeighborDensities(censdata, censdata2 = NULL, r = 20,
  plotdim = c(1000, 500), mindbh = 10, type = "count",
  include = c("A"))

Arguments

censdata
A table of plot census data.

censdata2
(NULL) an optional dataset for focal individuals, for example a subset of only one species or a list of seedling coordinates that are not part of census data. It must contain coordinates (gx,gy), species id (sp) and tag. It will count the neighbors in censdata but only for these focal coordinates.

r
Radius.

plotdim
The x and y dimensions of the plot.

mindbh
The minimum diameter above which the counts are done. Trees smaller than mindbh are excluded. If NULL, all living trees are included.

type
(count) calculates count of stems ('count') or sum of basal areas ('basal'). Note that for the sum of basal areas is better to use the plot stem table, otherwise only one stem per tree will be included .

include
(c("A")) a vector of status to include in the neighbors, by default it counts only alive individuals, for all status use c("A","P","D","M").

Value

A data.frame with two columns listing the density of consp and hetsp, and rows equal to the size of censdata if censdata2 is absent, or censdata2 if it is present.

Details

Arguments mindbh and r refer to neighbours

censdata should have the coordinates gx, gy, species id sp, status, tag and dbh.

Dependencies: CalcRingArea and circlearea from the CTFS package, and splancs package.

Examples

# NOT RUN {
# sum consp and hetsp neighbors for all stems in the plot  
neighbor.counts <- NeighborDensities(bciex::bci12t7mini, r=20, type='count')

# sum consp and hetsp neighbors for only one species: 
one.sp = subset(bciex::bci12t7mini, sp=="quaras")
neighbor.counts <- NeighborDensities(bciex::bci12t7mini, one.sp, type='count')
# }