Correct coordinates, making them from 0 to 20.

This function is called by fullplot.imageJ in the case that the maps are subquadrats to correct the coordinates and make them from 0 to 20 depending on what section of the map the subquadrat belongs to.

SectionCorrection(pts, subquad, gridsize, subquadsuffix)

Arguments

pts
the coordinates to be corrected

subquad
the subquadrat to be corrected

gridsize
Size of each individual map.

subquadratsuffix
Suffixes used to indicate what section of the map the subquadrat refers to. They should be named clockwise from the lower left. It can be used for map files that are smaller than 20x20m (i.e. 10x10m).