Collect abundances of all species across several censuses.

Collect abundances of all species across several censuses.

abund.manycensus(allcns = list(bci.full1, bci.full2, bci.full3), mindbh,
  dbhunit = "mm", type = "abund", excludespp = NULL,
  excludestatus = NULL)

Arguments

allcns
Full R census, submitted as a list (as many as desired).

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

dbhunit
'cm' or 'mm', only used for basal area

type
Either 'abund' (default) for abundance, 'ba' for basal area, or 'agb'.

excludespp
A character vector of species codes can be submitted as excludespp, for instance those for unidentified trees.

excludestatus
By default, any tree ever given code M is not counted in any census, but set excludestatus = NULL to include them.

Examples

not_run({
  N = abund.manycensus(
    allcns = list(
      bci::bci12full1, bci::bci12full2, bci::bci12full3, bci::bci12full4
    ),
    mindbh = 10,
    type = 'abund',
    excludespp = 'uniden',
    excludestatus = 'M'
  )
  head(N)
  colSums(N)
  apply(N,2,countspp)

  N = abund.manycensus(
    allcns = list(bci::bci12full5, bci::bci12full6),
    mindbh = 10,
    type = 'abund',
    excludespp = c('uniden', 'tremxx'),
    excludestatus = NULL
  )
})