abundance.spp

A wrapper to calculate total abundance (or ba or agb) for each species in given dbh categories. The dbh categories are set with dbhbreaks. See abundance() for description of the other arguments and return value.

abundance.spp(censdata, type = "abund", dbhunit = "mm",
  alivecode = c("A"), dbhbreaks = c(10, 100, 300))

Arguments

dbhbreaks
a vector of dbhs to define divisions of categories; the last category will be >= the final division