Abundance, basal area, or agb of every species by quadrat.

Finds abundance, basal area, or agb of every species per square quadrat of any size; plotdim is the x dimension then y dimension of the plot and must be set correctly; gridsize is the quadrat dimension. The plot is divided into a checkerboard of non-overlapping, space-filling squares.

abundanceperquad(censdata, mindbh = 10, plotdim = c(1000, 500),
  gridsize = 100, type = "abund", dbhunit = "mm")

Arguments

censdata
A table of plot census data.

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

plotdim
The x and y dimensions of the plot.

gridsize
Side of the square quadrate.

type
Either 'abund' (default) for abundance, 'ba' for basal area, or 'agb'.

dbhunit
'cm' or 'mm', only used for basal area

Value

See abundance().

Details

If the plot dimensions is not an exact multiple of the quadrat size, then a strip at the upper edge of the plot (north and east if plot is on cardinal directions) is omitted. For example, if gridsize = 40 and plotdim = 500, then there are an extra 20 meters at the upper boundary omitted from the calculations.

The array of abundances per quadrat is useful for similarity, counting species and stems per quadrat, etc.

See also

abundance()

Examples

not_run({
  Nperquad = abundanceperquad(
    bci::bci12full6,
    plotdim = c(1000, 500),
    gridsize = 100,
    type = 'abund'
  )
  colSums(Nperquad$abund)
  apply(Nperquad$abund, 2, countspp)
  plot(colSums(Nperquad$abund), apply / (Nperquad$abund, 2, countspp))
})

#> Error: <text>:10:56: unexpected ','
#> 9:   apply(Nperquad$abund, 2, countspp)
#> 10:   plot(colSums(Nperquad$abund), apply / (Nperquad$abund,
#>                                                            ^