Calculates the slope of all quadrats in a plot.

Calculates the slope of all quadrats in a plot.

allquadratslopes(elev, gridsize = 20, plotdim = c(1000, 500),
  edgecorrect = TRUE)

Arguments

elev
A list named col containing a data.frame with variables x, y and elev. This object is very specific; See example.

gridsize
Side of the square quadrate.

plotdim
The x and y dimensions of the plot.

Details

allquadratslopes() goes through all 20x20 m quadrats in a plot and finds the slope, mean elevation, and convexity of each. Convexity is the mean elevation of one 20x20 m quadrat relative (minus) the mean of its immediate neighbors.

Helene Muller-Landau added a section to correct convexity in edge quadrats.

See also

calcslope(), quadslope()

Examples

# NOT RUN {
# The input to elev is very specific; you may need to tweak it.
elev_asis <- bciex::bci_elevation
head(elev_asis)
elev_tweaked <- list(col = elev_asis)
head(elev_tweaked)

result <- allquadratslopes(
  elev = elev_tweaked,
  gridsize = 20,
  plotdim = c(1000, 500),
  edgecorrect = TRUE
)
head(result)
str(result)
# }