Reformat the output from demographic functions from list to dataframe.

Takes output of a demographic analysis (produced by functions growth, mortality, or pop.change) and converts into one dataframe. Only indicated dbh categories are included.

assemble.demography(output, type = "g", whichdbhcat = 1,
  date1 = "1960-1-1")

Arguments

output
The results from one of the demographic functions

type
A character. Either 'g' for growth, 'm' for mortality, 'ba' for basal area, 'agb' for biomass, 'r' for recruitment, or 'a' for abundance.

whichdbhcat
A number indicating how many dbh categories to include, cannot exceed the number of columns in output.

date1
date of first enumeration at site, format: yyyy-mm-dd.

Value

A dataframe.

Examples

# NOT RUN {
data = pop.change(
  bciex::bci12t5mini, bciex::bci12t6mini,
  split1 = bciex::bci12t5mini$sp
)
result = assemble.demography(data, type = 'a', whichdbhcat = 1)
data = growth(bciex::bci12t5mini, bciex::bci12t6mini, split1 = bciex::bci12t5mini$sp)
result = assemble.demography(data, type = 'g', whichdbhcat = 1)
data = mortality.eachspp(
  bciex::bci12t5mini, bciex::bci12t6mini,
  classbreak = c(10, 100)
)
result1 = assemble.demography(data, type = 'm', whichdbhcat = 1)
result2 = assemble.demography(data, type = 'm', whichdbhcat = 2)
# }