Returns the basal area summed over all submitted dbhs.

Returns the basal area summed over all submitted dbhs. NAs can be included, as sum will be completed with na.rm = TRUE.

basum(dbh, mindbh = 10, dbhunit = "mm")

Arguments

dbh
Stem diameter, usually at breast height (see data dictionary at http://ctfs.si.edu/Public/DataDict/data_dict.php).

mindbh
The minimum diameter above which the counts are done. Trees smaller than mindbh are excluded. If NULL, all living trees are included.

dbhunit
'cm' or 'mm', only used for basal area