Make a graph, 4 panels, of AGB growth and model fit of many species...

Make a graph, 4 panels, of AGB growth and model fit of many species overlaid, predicted functions only. Must submit names of four data objects (not the object, just the name in quote marks!).

binGraphManySpecies.Panel(fulldataname, sitename, darken = 8, xrange = c(-3,
  3), yrange = c(0, 0.1), export = pdf,
  outfile = "growth/linearbin.multi.pdf", h = 8, w = 10)

Arguments

export
File format to export to.

outfile
Filename for output.

h, w
A number giving the height and width.

Examples

# NOT RUN {
binGraphManySpecies.Panel(
  c(
    'linearbin.fit.allspp',
    'linearbin.fit.trim3',
    'linearbin.fit.edoro3',
    'linearbin.fit.lenda3'
  ),
  export = pdf,
  yrange = c(0, .25),
  sitename = c('Fushan', 'BCI', 'Edoro', 'Lenda'),
  darken = 5,
  xrange = c(-3, 3.5)
)
# }