Calculate change in biomass (agb).

Like growth(), but calculates change in biomass (agb) instead of dbh. The census tables must have a column called agb. There is no trimming done at all -- every tree is included, and its entire biomass (the agb column in the standard CTFS data object has total agb, all stems included.)

biomass.growth(census1, census2, rounddown = FALSE, stdev = FALSE,
  mindbh = 10, split1 = NULL, split2 = NULL)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

rounddown
If TRUE, all dbh < 55 are rounded down to the nearest multiple of 5.

stdev
Logical. Default (FALSE) returns confidence limits, otherwise returns the SD in growth rate per group.

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

split1
a vector of categories, one per individual

split2
another vector of categories, one per individual