Calculate flow using Seibert & McGlynn algorithm.

Calculate flow using Seibert & McGlynn algorithm.

calc.gradient(elev, grid = 20)

Arguments

elev
A 3x3 matrix of elevations; works on central point.

This calculates the gradient for the 8 triangular facets around the center point, following Seibert & McGlynn

The output is a data.frame of direction and slope for the 8 facets, starting with the lower left and moving clockwise

Examples

# NOT RUN {
data <- c(268.7, 275.9, 283.2, 275.9, 282.8, 290.0, 283.2, 290.0, 297)
z <- matrix(
  data,
  nrow = 3,
  byrow = TRUE
)
calc.gradient(z)
# }