complete.plotmap

This creates a map with every individual shown. The area to be mapped can be chosen, allowing maps of small areas.

On top of the map of every individual, individual species can be overlaid in different colors. If nospp is set NULL, no species are added. If spnames is set NULL, then the most abundant species in the plot are chosen, up to the number nospp.

complete.plotmap(cns = bci.full6, spnames = NULL, mindbh = 10,
  export = "no", nospp = 3, plotdim = c(1000, 500), clrlist = c("blue",
  "green", "red", "yellow", "gray"), ptsize = c(0.45, 0.3), xrange = c(0,
  100), yrange = c(0, 100), wd = 1100, ht = 850, side = 6,
  labsize = 1.75, axisdiv = 10, filepath = "/home/condit/data/maps/",
  outfile = "fullplotmap")

Arguments

cns:

a full census dataset (all species)

spnames:

names of species to map, using the mnemonic in the R tables

mindbh:

smallest dbh to include

export:

set to 'no'to graph to screen, 'pdf'to export to pdf (see define.graphwindow function)

nospp:

number of species to overlay; can be NULL or 0 for none

plotdim:

x and y plot dimensions

clrlist:

colors to use for the species to be overlaid

ptsize:

size of points, the first used for the background of all species, the second for the individual species

xrange:

minimum and maximum x coordinates of area graphed

yrange:

minimum and maximum y coordinates of area graphed

wd:

graph width; see map() function

ht:

graph height, same units as wd

side:

the side in inches of the graph; see map()

filepath:

folder to save output

outfile:

filename for output