A Mimimal Package Containing the Original CTFS R Package (Richard Condit)

ctfs is to be programatically imported by other packages developed by ForestGEO (http://www.forestgeo.si.edu/). If you need the original CTFS R Package, visit: http://ctfs.si.edu/Public/CTFSRPackage/; there you will find not only all functions but also their documentation.

Acknowledgement

The BCI forest dynamics research project was founded by S.P. Hubbell and R.B. Foster and is now managed by R. Condit, S. Lao, and R. Perez under the Center for Tropical Forest Science and the Smithsonian Tropical Research in Panama. Numerous organizations have provided funding, principally the U.S. National Science Foundation, and hundreds of field workers have contributed.

References

Anderson-Teixeira, K. J., Davies, S. J., Bennett, A. C., Gonzalez-Akre, E. B., Muller-Landau, H. C., Joseph Wright, S., . . . Zimmerman, J. 2015. CTFS-ForestGEO: a worldwide network monitoring forests in an era of global change. Global Change Biology, 21(2), 528-549. doi:10.1111/gcb.12712 Hubbell, S.P., Condit, R., and Foster, R.B. 2010. Barro Colorado Forest Census Plot Data. http://ctfs.si.edu/webatlas/datasets/bci.

Condit, R. 1998. Tropical Forest Census Plots. Springer-Verlag and R. G. Landes Company, Berlin, Germany, and Georgetown, Texas.

Hubbell, S.P., R.B. Foster, S.T. O'Brien, K.E. Harms, R. Condit, B. Wechsler, S.J. Wright, and S. Loo de Lao. 1999. Light gap disturbances, recruitment limitation, and tree diversity in a neotropical forest. Science 283: 554-557.