Run model.littleR.Gibbs for a series of census databases.

Run model.littleR.Gibbs() for a series of census databases, for every successive pair, then the first to the last. Then repeat for 10 times the initial mindbh.

fitSeveralAbundModel(allcns = list(bci.full1, bci.full2, bci.full3),
  sptable = bci.spptable, mindbh, excludespp = NULL,
  useIDlevel = TRUE, abundrange = c(1, 1e+06), start = c(-3, 0.8,
  0.01, -0.5), modeltype = "asympower",
  bad.modelparam = bad.asympower.param, steps = 10000, burn = 1000,
  show = 250, debug = FALSE)

Arguments

allcns
Full R census, submitted as a list (as many as desired).

mindbh
The minimum diameter above which the counts are done. Trees smaller than mindbh are excluded. If NULL, all living trees are included.

abundrange
the default includes every species, but this can be set to a minimum and maximum abundance (first census); species with abundances outside the range are excluded

start
A vector giving the starting set of parameters for the model. It must be as long as the number of parameters required by the model.

modeltype
Functional forms to fit to the distribution; it can be:

Gaussian modeltype = "norm", with the quotes

Asymmetric Gaussian (a different standard deviation on left and right of the mode) modeltype = "asymnorm", with the quotes

Laplace (exponential distribution, with mirror image for negative values) modeltype = "symexp", with the quotes

Asymmetric Laplace (different rate constant for left and right of the center) modeltype = "asymexp", with the quotes

Asymmetric power distribution (different rate constant for left and right of the center) `modeltype = "asympower", with the quotes.

bad.modelparam
name of a function which checks the model parameters for bad values; for modeltype asymexp, must be bad.asymexp.param, for modeltype asympower, must be bad.asympower.param

steps
The number of steps to run the Gibbs sampler.

burn
number of steps of sampler to exclude as burn-in

debug
Logical. If TRUE, call browser to debug.

See also