graph.abundmodel

Output histograms of little.r across species, observed and fitted, using the result of model.littleR.Gibbs. The histogram of black points is all species, blue points only those starting with N >= minabund. If the argument mortcorr=TRUE, a graph of mortality rate vs. population change for every species is also produced. Otherwise, a table of the species with biggest increases and biggest decreases in abundance is printed to the screen.

graph.abundmodel(fit, datafile = NULL, div = 0.01, tinydiv = 0.001,
  modeltype = "asympower", xrange = NULL, yrange = NULL, minabund = 50,
  conf = 25, returnextreme = TRUE, xname = "Rate of population change",
  yname = "Frequency", ltype = "solid", lwidth = 2, modelclr = "green",
  bartype = FALSE, addpts = TRUE, makeleg = TRUE, add = FALSE,
  newgraph = FALSE, ax = TRUE, mortcorr = FALSE, debug = FALSE)

Arguments

fit

result of model.littleR.Gibbs

datafile

optional name of file where the fitted result is saved

div

width of bins for histogram of observed rate of population change

tinydiv

width of bins used to draw the fitted distribution

modeltype

form of probability distribution, matching what was used when fit was created by model.littleR.Gibbs

xrange,

yrange range of graph's x-axis and and y-axis

minabund

minimum abundance of species to be used in histogram of observed rates of population change

conf

number of alternate fits to graph, as indication of confidence; if conf=NULL, no confidence lines are added

returnextreme

whether to print a list of the fastest increases and decreases in abundance to the screen

xname,

yname axis names

ltype,

lwidth, modelclr type, width, color of the line showing fitted distribution

bartype

if TRUE, histogram is bar graph

addpts

if TRUE, histogram is a point graph

makeleg

whether to add legend

add

just like all R graphs, whether to add points or line to an existing graph

ax

if FALSE, the axes are not added

mortcorr

whether to graph the correlation between mortality and population change across species

debug

if TRUE, call browser to debug

graphit

if set to false,