Pass the output of extract.growth with every individual's growth (f...

Pass the output of extract.growth with every individual's growth (full), and another after outliers have been trimmed. This finds the records trimmed and overlays them on the entire graph. If a model is submitted, then the curves are graphed too.

graph.outliers.spp(full, trimmed, spname = "gustsu", fit = NULL,
  size = "agb", export = NULL, xtitle = "log(agb)", ytitle = "growth")

Arguments

fit
A fitted model.

Examples

# NOT RUN {
full = extract.growthdata(
  census1 = bci::bci12full5,
  census2 = bci::bci12full6,
  growcol = 'incgr',
  growthfunc = growth.biomass.indiv,
  logit = 'x',
  rounddown = FALSE,
  mindbh = 100,
  dbhunit = 'mm',
  err.limit = 4000,
  maxgrow = 7500
)
trimmed = extract.growthdata(
  census1 = bci::bci12full5,
  census2 = bci::bci12full6,
  growcol = 'incgr',
  growthfunc = growth.biomass.indiv,
  logit = 'x',
  rounddown = FALSE,
  mindbh = 100,
  dbhunit = 'mm',
  err.limit = 4,
  maxgrow = 75
)
# }