Annual Growth Rates by Categories.

The principle growth function, constructed like recruitment and mortality. It requires two complete datasets, one per census, with dbh, pom, and date for every individual of all species in at least 2 censuses (see Data Format).

growth(census1, census2, rounddown = FALSE, method = "I",
  stdev = FALSE, dbhunit = "mm", mindbh = 10, growthcol = "dbh",
  err.limit = 4, maxgrow = 75, split1 = NULL, split2 = NULL)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

rounddown
If TRUE, all dbh < 55 are rounded down to the nearest multiple of 5.

method
Use 'I' to calculate annual dbh increment: (dbh2 - dbh1)/time, or 'E' to calculate the relative growth rate (log(dbh2) - log(dbh1))/time.

stdev
Logical. Default (FALSE) returns confidence limits, otherwise returns the SD in growth rate per group.

dbhunit
'cm' or 'mm', only used for basal area

mindbh
The minimum diameter above which the counts are done. Trees smaller than mindbh are excluded. If NULL, all living trees are included.

growthcol
defines how growth is measured, either 'dbh'or 'agb'(agb=biomass)

err.limit
A number. Numbers such as 10000 are high and will return all measures.

maxgrow
A number. Numbers such as 10000 are high and will return all measures.

split1
a vector of categories, one per individual

split2
another vector of categories, one per individual

Value

A list with components:

rate, the mean annualized growth rate per category selected, either dbh increment, or relative growth

N, the number of individuals included in the mean (not counting any excluded)

clim, width of confidence interval; add this number to the mean rate to get upper confidence limit, substract to get lower

dbhmean, mean dbh in census 1 of individuals included

time, mean time interval in years

date1, mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer)

date2, mean date in census 2

Pass the list to assemble.demography() (in utilities.r) with type = "g" to convert the list to a data.frame.

Details

It calculates the mean growth rate in one or more categories defined by the split variables, split1 and split2. The column date is required for annualizing rates.

The columns status and stemID are both required, in order to determine which stems should have dbh change calculated.

The function trim.growth() handles all checks for trees to include; excluded are (a) cases where the stemID changes, (b) extreme values based on err.limit and maxgrow, (c) and trees below a minimum dbh in the first census.

See also

Examples

# NOT RUN {
CTFSplot("bci", 56)
growth.data = growth(bciex::bci12t5mini, bciex::bci12t6mini)
growth.data$rate
growth.data = growth(
  bciex::bci12t5mini, bciex::bci12t6mini,
  split1 = bciex::bci12t5mini$sp
)
growth.data$rate
assemble.demography(grow.data, type = 'g')
# }