Fit a regression line through log growth against log dbh, binning on dbh.

Fitting a regression line through log growth against log dbh, binning on dbh.

growth.flexbin(growthtable, sizecol = "dbh", nobin = 2, start = NULL,
  startsd = NULL, sdmodel = linear.model.ctr, badsdfunc = NULL,
  method = "Gibbs", rep = 1200, show = 100, burn = 200, ...)

Arguments

growthtable
A table with 2 columns, one named size (ie, dbh), other growth.

method
If method = "Gibbs", runs the Gibbs sampler starting with a set of parameters found by stats::optim().

...
Conditionally passed to a number of functions; see function definition.

Details

Since stats::optim() fails if the initial likelihood is -Inf, it became necessary to check the parameters with bad.binparam() before running.