Table growth rate of every individual both relative and dbh-increment.

Table growth rate of every individual both relative and dbh-increment.

growth.indiv(census1, census2, rounddown = FALSE, mindbh = 10,
  dbhunit = "mm", err.limit = 4, maxgrow = 75)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

rounddown
If TRUE, all dbh < 55 are rounded down to the nearest multiple of 5.

mindbh
The minimum diameter above which the counts are done. Trees smaller than mindbh are excluded. If NULL, all living trees are included.

dbhunit
'cm' or 'mm', only used for basal area

err.limit
A number. Numbers such as 10000 are high and will return all measures.

maxgrow
A number. Numbers such as 10000 are high and will return all measures.

Value

This returns a complete table with growth rate of every individual, both relative and dbh-increment. The table also includes most other key pieces of information for every individual: species, dbh, coordinates. Growth is trimmed with trim.growth(), and growth is returned as NA if the individual is excluded; note, though, that every individual tree is always included in the table, even if growth = NA. Arguments are the same as in growth().

Examples

# NOT RUN {
growth.table = growth.indiv(bciex::bci12t5mini, bciex::bci12t6mini)
# }