Maps a response variable with Rs contour or image functions, or bot...

Maps a response variable with R's contour or image functions, or both, using the output of coldata.to.imagemat.

Also can return contours as vector data, using R's contourLines. The graph can be sent to the screen or exported.

There are many options for tweaking the graph, explained with the arguments.

imageGraph(matrixdata, levels = NULL, breaks = c(-33000, 0, 150, 250, 500,
  1000, 2000), xname = "x", yname = "y", xrange = NULL, yrange = NULL,
  axisdiv = c(100, 100), img = TRUE, cntrs = FALSE, filled = TRUE,
  drawgrid = FALSE, addaxes = 1:4, clrs = NULL, cntcolor = "black",
  lwidth = 1, returnline = FALSE, newgraph = FALSE, add = FALSE,
  plotsize = 6, h = 11, w = 11, export = X11, graphfile = NULL)

Arguments

matrixdata
a list having components x, y, and mat, exactly the output of coldata.to.imagemat

levels
if set, then contours are drawn (or colors assigned) to breaks are at levels equally-spaced divisions of the response variable; can be NULL to let breaks control

breaks
if set, a vector of breaks for the image colors and the contour lines; if NULL, levels controls the breaks

xname, yname
names for the axes

xrange
if NULL, all x are included, otherwise, the graph will only include x inside xrange

yrange
likewise for y

axisdiv
the division between tick marks; must be a vector of 2, first for x ticks then for y ticks

img
TRUE or FALSE, whether to draw the color image

cntrs
TRUE or FALSE, whether to add contour lines; both img and cntrs can be TRUE

filled
TRUE or FALSE, whether to draw filled contours; if TRUE, neither image or standard contours are included

clrs
colors for the image; if NULL, they are chosen by default, otherwise, must be 1 + the number of breaks

returnline
TRUE or FALSE, whether to return the contour lines; see R's base function contourLines for details

newgraph
if FALSE, just graph to the screen and export is ignored; use TRUE to create a new plot window or export to a file

plotsize
size of graph's x-dimension in inches; ignored if newgraph=FALSE

h, w
height and width of graph window in units which depend on the export device chosen (some are pixels, some are inches); ignored if newgraph=FALSE

export
if newgraph=TRUE, this defines the graphics device; options include X11, quartz, win.graph, win.metafile, png, bmp, jpeg, pdf (not quoted!)

graphfile
the file name if export is pdf, png, jpeg, win.metafile, etc.

Examples

not_run({

  CTFSplot('bci', 6, 'full')

  N = colSums(
    abundanceperquad(bci.full6, plotdim = c(1000, 500), grid = 20)$abund
  )
  matdata = coldata.to.imagemat(z = N,
    gridsize = 20,
    plotmax = c(1000, 500))
  imageGraph(
    matrixdata = matdata,
    breaks = seq(100, 400, by = 25),
    img = TRUE,
    cntrs = FALSE,
    newgraph = FALSE,
    plotsize = 6,
    h = 11,
    w = 11
  )
  imageGraph(
    matrixdata = matdata,
    breaks = seq(100, 400, by = 25),
    img = TRUE,
    cntrs = TRUE,
    newgraph = FALSE,
    plotsize = 6,
    h = 11,
    w = 11
  )
  imageGraph(
    matrixdata = matdata,
    breaks = seq(100, 400, by = 25),
    img = TRUE,
    cntrs = FALSE,
    newgraph = TRUE,
    export = win.graph,
    h = 11,
    w = 11,
    plotsize = 8
  )
})