imageJ.to.lxly

Convert map pixels from digitzing program to plot coordinates. Uses 2 functions from geometry.r: pts.to.interceptslope=function(pt1,pt2) perpendicular.distance=function(b,m,x,y) return(sqrt((y-m*x-b)^2/(1+m^2))) intersection.of.lines=function(b1,m1,b2,m2)

imageJ.to.lxly(textfile = samplemapfile, delim = ",", lowerleft = "P2",
  upperleft = "P1", upperright = "P4", lowerright = "P3",
  gridsize = c(20, 20))

Arguments

textfile

the complete name of the textfile to convert, including the path

delim

the delimiter used to separate the fields in the map files

lowerleft,

upperleft, upperright, lowerright the tags in each of the map files with the calibrated corners

gridsize

size of each individual map