individual_grow.table

Create a table of individual trees and their growth over two censuses, with many species included. The option rnd is used to rounddown dbhs for certain intervals. The flag ctr is used to center the time variable, which is the number of years since 1960 of the interval midpoint. There is a logarithmic transformation, for which all growth<=0 is converted to mingrow. There is also a power transformation, where each growth rate is raised to the power given by powertransformation. In the latter, negative growths are transformed to negative, so do not need to be corrected.

individual_grow.table(cnsdata = list(bci.full1, bci.full2, bci.full3,
  bci.full4, bci.full5, bci.full6, bci.full7), powertransformation = 0.45,
  rnd = c(FALSE, TRUE, rep(FALSE, 4)), mingrow = 0.1, mindbh = 10,
  maxdbh = 10000, maxerrorSD = 4, maxerrorGrow = 75, center = 1992,
  debug = FALSE)