Table individual trees and their survival status over two censuses.

Create a table of individual trees and their survival status over two censuses, with many species included.

individual_mort.table(cnsdata, mindbh = 10, maxdbh = 10000,
  alivecode = c("A", "AB", "AS"), center = 1992)

Arguments

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

alivecode
character, codes of the variable `status` that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

See also

individual_grow.table()