Map a single species.

This does the mapping for a single species, called from map() but also useful on its own.

map1species(sppdata, plotdim = c(1000, 500), xrange = c(0, plotdim[1]),
  yrange = c(0, plotdim[2]), color = "black", bgcolor = "white",
  axspos = 0.5, xaxis = TRUE, yaxis = TRUE, symbol = 16,
  size = rep(-1, 3), cutoff = c(10, 100, 300, 3000), axisdiv = 100,
  labsize = 1.75, plotside = 4, add = FALSE)

Arguments

plotdim
The x and y dimensions of the plot.

xrange
Minimum and maximum x and y coordinates of area graphed. If `NULL`, all `x` (or `y`) are included.

yrange
Minimum and maximum x and y coordinates of area graphed. If `NULL`, all `x` (or `y`) are included.

bgcolor
The background color. Defaults to white. For presentation exports, try bgcolor="transparent".

axspos
Distance between axis numbers and axis.

size
The size of plotting points, to match the number of diameter breaks. If NULL, a default set is assigned. This can require fiddling, as big points do not work for really abundant species, and small points for rare species.

cutoff
Diameter breaks for changing size of plotting points.

axisdiv
the division between tick marks; must be a vector of 2, first for x ticks then for y ticks

labsize
Size of axis labels.

plotside
The vertical dimension (inches) of the map.

add
Logical. Whether to add elements such as points or line to an existing graph. if TRUE, there must be an appropriate existing graph, otherwise, a new graph is drawn.

See also

map().