mortality

Mortality is the main function, and is constructed like growth and recruitment. It requires two complete datasets, one per census, with dbh, pom, and date for every individual of all species in at least 2 censuses (see Data Format). It can then calculate mortality for the entire forest, or based on one or two user-submitted factors. Mortality is based on only on the column status: any tree without an alivecode in census 2 is considered dead. Individuals whose status is NA in either census are deleted from all counts, since it's impossible to count them either as survivors or dead. It requires fill.dimension and climits in utilities.r. Output of the mortality function is a list with components: <li>N, the number of individuals alive in the census 1 per category selected <li>D, the number of individuals no longer alive in census 2 <li>rate, the mean annualized mortality rate constant per category selected, calculated as (log(N)-log(S))/time <li>upper, upper confidence limit of mean rate <li>lower, lower confidence limit of mean rate <li>time, mean time interval in years <li>date1, mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer) <li>date2, mean date in census 2 <li>dbhmean, mean dbh in census 1 of individuals included

Pass the list to assemble.demography (in utilities.r) with type="m" to convert the list a data.frame.

mortality(census1, census2, alivecode = c("A", "AB", "AS"), split1 = NULL,
  split2 = NULL)

Arguments

split1

and split2 must both be vectors of character variables with exactly as many elements as there are rows in the tables census1 and census2 (or both can be NULL), for instance, species names, dbh categories, or quadrat numbers

Generally,

alivecode="A" suffices, as this is the standard in CTFS data for a living tree; "AS" and "AB" are seldom used now

Examples

## Not run: ------------------------------------
# CTFSplot("bci",56)
# mort.data=mortality(bci.full5,bci.full6)
# mort.data$rate
# mort.data=growth(bci.full5,bci.full6,split1=bci.full5$sp)
# mort.data$rate
# assemble.demography(mort.data,type='m')
## ---------------------------------------------