Calculate mortality for each species in given dbh categories.

Calculate mortality for each species in given dbh categories. It sets the split variables using the species name and submitted dbh classbreaks and then uses mortality to do the calculation.

mortality.eachspp(census1, census2, classbreak = c(10, 100, 300),
  alivecode = c("A", "AB", "AS"))

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

classbreak
xxxdocparam

alivecode
character, codes of the variable status that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

Value

The list from mortality(), which can be passed to assemble.demography() for a convenient format.

Examples

# NOT RUN {
CTFSplot("bci", 5:6)
mort.data = mortality.eachspp(bciex::bci12t5mini, bciex::bci12t6mini)
mort.table1 = assemble.demography(mort.data, type = "m", whichdbhcat = 1)
mort.table2 = assemble.demography(mort.data, type = "m", whichdbhcat = 2)
mort.table3 = assemble.demography(mort.data, type = "m", whichdbhcat = 3)
# }