Show model fits for 1, 2, 3, and 4 bins on each species.

Show model fits for 1, 2, 3, and 4 bins on each species.

overlay.growthbinmodel(fit, bins = 1:4, regclr = "green",
  modelclr = "blue", graphdiv = 15, add = FALSE, newgraph = TRUE,
  export = pdf, outfile = "growth/linearbin.overlay.pdf", h = 8, w = 10)

Arguments

fit
A fitted model.

add
Logical. Whether to add elements such as points or line to an existing graph. if TRUE, there must be an appropriate existing graph, otherwise, a new graph is drawn.

newgraph
if FALSE, just graph to the screen and export is ignored; use TRUE to create a new plot window or export to a file.

outfile
Filename for output.

h, w
A number giving the height and width.