Export species maps as png.

Export species maps as png. There will be one for every species chosen. See pdf.allplot.

png.allplot(splitdata = bci.split6, spplist = plotspp$bci,
  elev = ctfs.elev$bci$mat, plotdim = c(1000, 500), topoint = 2,
  plotside = 9, h = 850, w = 1100, ptsizes = c(0.3, 0.45, 0.65),
  dbhcut = c(10, 100, 300, 3000), dead = FALSE, export = TRUE,
  path = "/home/condit/data/maps/bci/")

Arguments

spplist
A table with species codes and Latin names can be submitted as well, so that the full genus-species is added to plot. This must have species codes as row names. It should be the CTFS R format species table (eg, bci::bci12spptable). It can be set to NULL if not available, then only the species code (as submitted) appears on the map.

plotdim
The x and y dimensions of the plot.

topoint
Interval for topolines, if elevdata are submitted.

plotside
The vertical dimension (inches) of the map.

h, w
A number giving the height and width.

export
`export = "no"`, does nothing, so the next graph uses the default R device. With this option, the size of the graph cannot have been set larger than the default size, which is usually 7 inches. `export = "unix"`, a new graphics window in unix, size `h` and width `w`, will be opened allowing you to alter the size of the output. `export = "windows"`, or `export = "mac"` are similar to `export = "unix"` but for windows and mac. `export = format` exports the graph to a specific file format: "pdf", "png", "jpg", "bmp", "emf".

See also

map().