Convert quadrat names into x-y coordinates.

Convert quadrat names into x-y coordinates, assuming the first 2 digits are the column and the second two the row. Quad is a character.

quad.to.gxgy(quad, gridsize = 20, start = 0)

Arguments

quad
A character giving the quadrat name to convert to x-y coordinates.

gridsize
Side of the square quadrate.

Argument details

If the first row and column are 00, set start = 0, etc.