Find best fits for linearmodel.bin, with one set of data and a seri...

Find best fits for linearmodel.bin, with one set of data and a series of bins. This starts fit for a bin based on the best fit for the prior bin, thus always assuring an improved fit. This works best if the first bin=1, which is just a linear model and easily fit by optim and the Gibbs sampler.

run.growthfit.bin(growthdata, size = "dbh", startpar = c(0.03, 0.005),
  startsdpar = c(0.04, 0), sdmodel = linear.model.ctr, badsdfunc = NULL,
  binoption = 1:4, noreps = 5000, noburn = 2500, noshow = 500, ...)

Arguments

...
Arguments passed to growth.flexbin().