Likelihood function for a species mean (a scalar, one species at a ...

Likelihood function for a species mean (a scalar, one species at a time), given logMu and logSD and the data, N and S (just one species here, so all parameters are scalars). In the abundance model, the mortality parameter is involved in the likelihood for population change, and it must thus depend on the the fitted little r. Only used as a subroutine of the main modeling function, model.littleR.Gibbs.

spmean.mort.abundGibbs(spmean.mu, spmean.r, N1, S, time, MuSD, Rsd, type)

See also

spmean.mort.abundGibbs()