Summarize calculations based on the full Gibbs sampler.

Make summary calculations based on the full Gibbs sampler.

summaryMCMC(fit, model, error, sdmodel, badparam, paramfile = NULL,
  returnfull = TRUE, ...)

summaryModelMCMC(fit, model, error, sdmodel, badparam, paramfile = NULL,
  returnfull = TRUE, ...)

Arguments

fit
An object holding all steps of the sampler, plus data, observations, and likelihood.

badparam
The name of a function (unquoted) that tests a set of model parameters for validity; must return TRUE if parameters are valid, otherwise FALSE.

paramfile
If parameters were saved along the way to a text file, then this argument is used to name the file and restore them into a 3D array.

returnfull
If TRUE, the entire 3D array of parameters is also returned. Full likelihood at the best parameters is calculated and likelihood at each step in sampler are used to calculate DIC.

...
Arguments passed to llike.model.lmer() or full.likelihood.lmerBayes(), which may be problematic (xxx_amend_source in summaryMCMC()).

Value

Estimates of confidence limits of all parameters are returned.

Functions

summaryModelMCMC: See summaryMCMC().