Exclude growth rates

This is where growth rates are excluded. It is based on a linear model estimating the standard deviation of dbh measures (due to error, that is); the parameters slope and intercept define the linear relationship between this error deviation and dbh. Any case where the second dbh measure is more than 4 standard deviations below the first is marked false, meaning it will be excluded from growth measurements. The default values of slope and intercept are based on dbh remeasure tests done in both 1995 and 2000 at BCI. A line was fitted through the absolute dbh errors as a function of dbh in both years; the average slope and intercept is used here. The function also excludes any growth rate > 75 mm per yr, cases where the stemID changes, and if the POM changes by more than 5

trim.growth(cens1, cens2, time, slope = 0.006214, intercept = 0.9036,
  err.limit = 4, maxgrow = 75, pomcut = 0.05, mindbh = 10,
  dbhunit = "mm", exclude.stem.change = TRUE)

Arguments

cens1, cens2
xxxdocparam See census1 and census2 in biomass.change().

time
xxxdocparam Years between censuses.

slope
Slope of error of measurement line.

intercept
Intercept of error of measurement line.

err.limit
A number. Numbers such as 10000 are high and will return all measures.

maxgrow
A number. Numbers such as 10000 are high and will return all measures.

pomcut
A number. To include POM changes, set it to a high number, such as 10.

mindbh
The minimum diameter above which the counts are done. Trees smaller than mindbh are excluded. If NULL, all living trees are included.

dbhunit
'cm' or 'mm', only used for basal area

exclude.stem.change
Logical. FALSE includes cases where stemID changed, regardless of growth (it does not make sense to exclude a record based on growth when the growth is based on different stems).

Details

This function is usually only used inside the other growth functions. Note that trees are exclude if cens1$dbh < mindbh, but not if cens2$dbh < mindbh. All parameters for excluding growth measures based on error can be adjusted.