Wavelet variance curves for all species in one plot.

Function to calculate the wavelet variance curves for all species in one plot using quadrat indexation from CTFS package.

wavelet.allsp(censdata, plotdim = c(1000, 500), gridsize = 2.5,
  mindbh = NULL)

Arguments

censdata
A table of plot census data.

plotdim
The x and y dimensions of the plot.

gridsize
Size of the quadrats for the rasterization

mindbh
If analysis is to be done at different size classes.

Value

A matrix with:

scale: vector with scale of the analysis in meters;

variance: matrix with the normalized variance at each scale (columns) for each species (rows); density: matrix with the density per area and abundance for each species in the plot;

gridsize: grid size parameter used;

UCL: vector with the upper confidence limit for the null hypothesis;

LCL: vector with the lower confidence limit for the null hypothesis.

Arguments details

censdata: Must contain the variables gx, gy, dbh, status, and sp code.

Examples

not_run({
  wavelet.variances = wavelet.allsp(
    censdata = bci::bci12full1,
    plotdim = c(1000, 500)
  )
})