Create a dummy wsgdata data set

Create a dummy wsgdata data set with random values of species density.

wsgdata_dummy(censdata, plot = "bci")

Arguments

censdata
Census data, e.g. bciex::bci12t1mini.

plot
Character string giving the name of the plot where the data comes from (as it appears in the names of the R Analytical Tables).

Value

A data frame with one row per species in censdata and 3 variables.

Details

If you lack a real wood density table, you can use this function to create a dummy one. For example, this may help you to test biomass.CTFSdb() and density_ind().

See ?density_ind(), particularly notice the following: "There must be a table of wood density submitted (wsgdata), and this table must have a column sp with species names, a column plot, plus the wood density in a column called wsg (though the name of that column can be changed using the argument denscol)".

Examples

wsgdata_dummy(bciex::bci12s7mini)

#>         sp plot           wsg
#> 1   apeime  bci  2.419707e-01
#> 2   alchco  bci  5.399097e-02
#> 3   quaras  bci  4.431848e-03
#> 4   beilpe  bci  1.338302e-04
#> 5   luehse  bci  1.486720e-06
#> 6   casear  bci  6.075883e-09
#> 7   simaam  bci  9.134720e-12
#> 8   alsebl  bci  5.052271e-15
#> 9   pri2co  bci  1.027977e-18
#> 10  poular  bci  7.694599e-23
#> 11  hybapr  bci  2.118819e-27
#> 12  faraoc  bci  2.146384e-32
#> 13  cappfr  bci  7.998828e-38
#> 14  ingaum  bci  1.096607e-43
#> 15  tri2tu  bci  5.530710e-50
#> 16  picrla  bci  1.026163e-56
#> 17  guargu  bci  7.004182e-64
#> 18  guatdu  bci  1.758750e-71
#> 19  drypst  bci  1.624636e-79
#> 20  tet2pa  bci  5.520948e-88
#> 21  cassel  bci  6.902029e-97
#> 22  alibed  bci 3.174282e-106
#> 23  psycg3  bci 5.370560e-116
#> 24  ingape  bci 3.342714e-126
#> 25  protpa  bci 7.653930e-137
#> 26  chr2ar  bci 6.447260e-148
#> 27  pterro  bci 1.997889e-159
#> 28  ouralu  bci 2.277577e-171
#> 29  quasam  bci 9.551695e-184
#> 30  olmeas  bci 1.473646e-196
#> 31  eugeoe  bci 8.363952e-210
#> 32  swars2  bci 1.746366e-223
#> 33  ormocr  bci 1.341420e-237
#> 34  poutre  bci 3.790526e-252
#> 35  chr1ec  bci 3.940396e-267
#> 36  swars1  bci 1.506905e-282
#> 37  andiin  bci 2.120007e-298
#> 38  acaldi  bci 1.097221e-314
#> 39  mourmy  bci  0.000000e+00
#> 40  psycho  bci  0.000000e+00
#> 41  socrex  bci  0.000000e+00
#> 42  des2pa  bci  0.000000e+00
#> 43  malmsp  bci  0.000000e+00
#> 44  calolo  bci  0.000000e+00
#> 45  cordbi  bci  0.000000e+00
#> 46  loncla  bci  0.000000e+00
#> 47  allops  bci  0.000000e+00
#> 48  pipeco  bci  0.000000e+00
#> 49  laciag  bci  0.000000e+00
#> 50  hirttr  bci  0.000000e+00
#> 51  cupasy  bci  0.000000e+00
#> 52  heisco  bci  0.000000e+00
#> 53  tachve  bci  0.000000e+00
#> 54  ingago  bci  0.000000e+00
#> 55  gar2in  bci  0.000000e+00
#> 56  nects1  bci  0.000000e+00
#> 57  virose  bci  0.000000e+00
#> 58  ocotwh  bci  0.000000e+00
#> 59  turpoc  bci  0.000000e+00
#> 60  ingasa  bci  0.000000e+00
#> 61  protco  bci  0.000000e+00
#> 62  xyl1ma  bci  0.000000e+00
#> 63  ingaqu  bci  0.000000e+00
#> 64  cordla  bci  0.000000e+00
#> 65  protte  bci  0.000000e+00
#> 66  eugeco  bci  0.000000e+00
#> 67  maquco  bci  0.000000e+00
#> 68  aspicr  bci  0.000000e+00
#> 69  bactc1  bci  0.000000e+00
#> 70  sipapa  bci  0.000000e+00
#> 71  randar  bci  0.000000e+00
#> 72  dendar  bci  0.000000e+00
#> 73  soroaf  bci  0.000000e+00
#> 74  guarsp  bci  0.000000e+00
#> 75  licapl  bci  0.000000e+00
#> 76  coccma  bci  0.000000e+00
#> 77  tri2pa  bci  0.000000e+00
#> 78  gustsu  bci  0.000000e+00
#> 79  talipr  bci  0.000000e+00
#> 80  rinosy  bci  0.000000e+00
#> 81  guetfo  bci  0.000000e+00
#> 82  hameax  bci  0.000000e+00
#> 83  oenoma  bci  0.000000e+00
#> 84  zuelgu  bci  0.000000e+00
#> 85  eugene  bci  0.000000e+00
#> 86  cecrin  bci  0.000000e+00
#> 87  ingami  bci  0.000000e+00
#> 88  guapst  bci  0.000000e+00
#> 89  tab2ar  bci  0.000000e+00
#> 90  hassfl  bci  0.000000e+00
#> 91  cou2cu  bci  0.000000e+00
#> 92  unonpi  bci  0.000000e+00
#> 93  caseac  bci  0.000000e+00
#> 94  brosal  bci  0.000000e+00
#> 95  micone  bci  0.000000e+00
#> 96  ingas1  bci  0.000000e+00
#> 97  micoaf  bci  0.000000e+00
#> 98  crotbi  bci  0.000000e+00
#> 99  coccco  bci  0.000000e+00
#> 100 bactc2  bci  0.000000e+00
#> 101 celtsc  bci  0.000000e+00
#> 102 annoac  bci  0.000000e+00
#> 103 sipagu  bci  0.000000e+00
#> 104 ocotpu  bci  0.000000e+00
#> 105 sponmo  bci  0.000000e+00
#> 106 virosu  bci  0.000000e+00
#> 107 sloate  bci  0.000000e+00
#> 108 ocotce  bci  0.000000e+00
#> 109 anaxpa  bci  0.000000e+00
#> 110 dio2ar  bci  0.000000e+00
#> 111 pipeae  bci  0.000000e+00
#> 112 ingapu  bci  0.000000e+00
#> 113 tropra  bci  0.000000e+00
#> 114 sympgl  bci  0.000000e+00
#> 115 paligu  bci  0.000000e+00
#> 116 psycma  bci  0.000000e+00
#> 117 aegipa  bci  0.000000e+00
#> 118 stylst  bci  0.000000e+00
#> 119 ingath  bci  0.000000e+00
#> 120 ingafa  bci  0.000000e+00
#> 121 pentma  bci  0.000000e+00
#> 122 herrpu  bci  0.000000e+00
#> 123 vochfe  bci  0.000000e+00
#> 124 casesy  bci  0.000000e+00
#> 125 cedrod  bci  0.000000e+00
#> 126 ery2ma  bci  0.000000e+00
#> 127 ocotob  bci  0.000000e+00
#> 128 urerba  bci  0.000000e+00
#> 129 ingama  bci  0.000000e+00
#> 130 laetth  bci  0.000000e+00
#> 131 psycac  bci  0.000000e+00
#> 132 eugega  bci  0.000000e+00
#> 133 micoar  bci  0.000000e+00
#> 134 annosp  bci  0.000000e+00
#> 135 cecrob  bci  0.000000e+00
#> 136 sapiau  bci  0.000000e+00
#> 137 psycde  bci  0.000000e+00
#> 138 cordal  bci  0.000000e+00
#> 139 nectgl  bci  0.000000e+00
#> 140 pla1pi  bci  0.000000e+00
#> 141 cha2sc  bci  0.000000e+00
#> 142 sennda  bci  0.000000e+00
#> 143 pse1se  bci  0.000000e+00
#> 144 hampap  bci  0.000000e+00