Above ground biomass (agb) based on one of the Chave (Oecologia, 2005).

Compute biomass (agb) based on one of the Chave (Oecologia, 2005) models for tropical forest types.

AGB.ind(df, dbhunit = "mm", plot = "bci", wsgdata = wsg.ctfs2,
  forest = "moist", ht.param = NULL, htmodel = predht.asym)

Arguments

df
A table with dbh and species names

dbhunit
'cm' or 'mm', only used for basal area

plot
Character string giving the name of the plot where the data comes from (as it appears in the names of the R Analytical Tables).

wsgdata
A wood-density table, see density_ind() and wsgdata_dummy.

forest
A forest type (for most lowland tropical plots, the default "moist" is recommended.)

ht.param
Height parameters to calculate height from diameter for every tree, using the Chave model with height passed to htmodel. If NULL (defauld) AGB.ind() uses Chave equations without height.

htmodel
Height model to use with a non-NULL ht.param.

Value

A vector of biomass in kg for every individual in the submitted table df.

See also

density_ind(); Chave (Oecologia, 2005).

Examples

not_run({
  biomass <- AGB.ind(df = bci::bci12full1)
  hist(log(biomass), breaks = 100)
  sum(biomass, na.rm = TRUE) / 50
})