AGB of each tree, grouping all stems of one tree and adding there agbs.

Computes AGB of each tree in a table, grouping all stems of one tree and adding there agbs.

AGB.tree(df, dbhunit = "mm", plot = "bci", wsgdata = wsg.ctfs2,
  forest = "moist", ht.param = NULL, htmodel = predht.asym)

Arguments

df
The submitted table. Must have dbh, species name (sp), and a treeID to identify which tree every stem belong to. There must be just one dbh for each stem.

dbhunit
'cm' or 'mm', only used for basal area

plot
Character string giving the name of the plot where the data comes from (as it appears in the names of the R Analytical Tables).

Value

Adataframe with one row per tree, including the treeID and total agb per tree. Note that it will have fewer rows than the table submitted.

See also

This is called by biomass.CTFSdb() in the standard calculation of biomass for CTFS R tables.

Examples

# NOT RUN {
biomasstbl = AGB.tree(df = bci::bci12stem1)
dim(bci::bci12stem1)
dim(biomasstbl)
head(biomasstbl)
# }