The Chave 2005 Oecologia model for calculating biomass from dbh in ...

The Chave 2005 Oecologia model for calculating biomass from dbh in cm. All dbhs are submitted as a vector, and a vector of wood density of the same length must also be submitted (or a single wood density can be passed, to be used for every tree).

Parameter values for the 3 forest types according to Chave 2005 are hard-coded in the function.

The recommended CTFS use is with htparam=NULL, so height is not used. If height parameters and a height model are passed, then the height of every tree is calculated, and the Chave AGB formula that includes height is used. The default height parameters are from Chave et al 2003 on BCI biomass, and the default height function is predht.asym, provided in this file. But any height model can be substituted, providing the function name is passed and the necessary number of parameters included as htparam.

Returns a vector of biomass of same length as vector of dbh submitted.

This is called by AGB.tree in the standard calculation of biomass for CTFS R tables.

Chave.AGB(dbh, density = 0.62, htparam = c(41.7, 0.057, 0.748),
  heightmodel = predht.asym, forest = "moist")

Examples

not_run({
  testdbh=c(1,2,5,10,20,30,50,100,200)

  AGBmoist=Chave.AGB(dbh=testdbh,forest="moist")

  AGBwet=Chave.AGB(dbh=testdbh,forest="wet")

  plot(testdbh,AGBmoist,col="green",type="l")

  lines(testdbh,AGBwet,col="blue")
})