Abundance or basal area dividing data with 1 or 2 categorical variables

Calculates total abundance or basal area, dividing data with 1 or 2 categorical variables.

abundance(censdata, type = "abund", alivecode = c("A"), mindbh = NULL,
  dbhunit = "mm", split1 = NULL, split2 = NULL)

Arguments

censdata
A table of plot census data.

type
Either 'abund' (default) for abundance, 'ba' for basal area, or 'agb'.

alivecode
character, codes of the variable `status` that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

dbhunit
'cm' or 'mm', only used for basal area

split1
a vector of categories, one per individual

split2
another vector of categories, one per individual

Value

a named list of two elements, either abund or ba, and meandate (mean measurement date for all individuals in each category). Each element of the list is an array of one or two dimensions, depending on how many split variables were submitted: each of the dimensions of the array handles one of the categorical variables.

Details

The categorical variables must be submitted as vectors whose length matches exactly as the number of rows in the plot data submitted (so one per tree or stem). The first vector should be the one with the most categories (for instances, split1 = species, split2 = dbhcategory).

For basal area, pass a stem table to censdata; For abundance, use either the stem or full table to count stems or trees, respectively.

Examples

not_run({

  CTFSplot('bci',5:6,'full')

  CTFSplot('bci',5:6,'stem')
  total=abundance(bci::bci12full5,mindbh=10)
  total$abund
  total$meandate
  totalstem=abundance(bci::bci12stem5,mindbh=10)

  BAperSpecies=abundance(bci::bci12stem5,type='ba',mindbh=10,split1=bci::bci12stem5$sp)
  head(BAperSpecies$ba)
  head(BAperSpecies$meandate)
})