Tools for the Analysis of Forest Dynamics

This package provides tools for the analysis of forest dynamics. It is a slightly modifyed version of the original CTFS R Package by Richard Condit, available at http://ctfs.si.edu/Public/CTFSRPackage/. The minor changaes only aimed to transform the origial CTFSR Package into an R package sensu stricto; Although the name and documentation of a few functions slightly changed, the source code remains the same.

Acknowledgement

The BCI forest dynamics research project was founded by S.P. Hubbell and R.B. Foster and is now managed by R. Condit, S. Lao, and R. Perez under the Center for Tropical Forest Science and the Smithsonian Tropical Research in Panama. Numerous organizations have provided funding, principally the U.S. National Science Foundation, and hundreds of field workers have contributed.

References

Anderson-Teixeira, K. J., Davies, S. J., Bennett, A. C., Gonzalez-Akre, E. B., Muller-Landau, H. C., Joseph Wright, S., . . . Zimmerman, J. 2015. CTFS-ForestGEO: a worldwide network monitoring forests in an era of global change. Global Change Biology, 21(2), 528-549. doi:10.1111/gcb.12712 Hubbell, S.P., Condit, R., and Foster, R.B. 2010. Barro Colorado Forest Census Plot Data. URL.

Condit, R. 1998. Tropical Forest Census Plots. Springer-Verlag and R. G. Landes Company, Berlin, Germany, and Georgetown, Texas.

Hubbell, S.P., R.B. Foster, S.T. O'Brien, K.E. Harms, R. Condit, B. Wechsler, S.J. Wright, and S. Loo de Lao. 1999. Light gap disturbances, recruitment limitation, and tree diversity in a neotropical forest. Science 283: 554-557.