Mortality, forest-wide or based on one or two categories.

Calculate mortality for the entire forest, or based on one or two user-submitted factors.

mortality(census1, census2, alivecode = c("A", "AB", "AS"), split1 = NULL,
  split2 = NULL)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

alivecode
character, codes of the variable `status` that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

split1
a vector of categories, one per individual

split2
another vector of categories, one per individual

Value

Output of the mortality function is a list with components:

N, the number of individuals alive in the census 1 per category selected

D, the number of individuals no longer alive in census 2

rate, the mean annualized mortality rate constant per category selected, calculated as (log(N)-log(S))/time

upper, upper confidence limit of mean rate

lower, lower confidence limit of mean rate

time, mean time interval in years

date1, mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer)

date2, mean date in census 2

dbhmean, mean dbh in census 1 of individuals included

Pass the list to assemble.demography() with type = "m" to convert the list a data.frame.

Details

Mortality is the main function, and is constructed like growth and recruitment. It requires two complete datasets, one per census, with dbh, pom, and date for every individual of all species in at least 2 censuses (see Data Format).

Mortality is based on only on the column status: any tree without an alivecode in census 2 is considered dead. Individuals whose status is NA in either census are deleted from all counts, since it's impossible to count them either as survivors or dead.

Examples

# NOT RUN {
CTFSplot("bci", 56)
mort.data = mortality(bci::bci12full5, bci::bci12full6)
mort.data$rate
mort.data = growth(
  bci::bci12full5, bci::bci12full6,
  split1 = bci::bci12full5$sp
)
mort.data$rate
assemble.demography(mort.data, type = 'm')
# }