Change in abundance, basal area, or agb in two censuses.

Finds abundance, basal area, or agb in two censuses and the rate of change between them. Accepts two dataframes, each an R Analytical Table for one census, the earlier census first.

pop.change(census1, census2, type = "abund", dbhunit = "mm",
  alivecode = c("A"), mindbh = NULL, split1 = NULL, split2 = NULL)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

type
Either 'abund' (default) for abundance, 'ba' for basal area, or 'agb'.

dbhunit
'cm' or 'mm', only used for basal area

alivecode
character, codes of the variable `status` that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

split1
a vector of categories, one per individual

split2
another vector of categories, one per individual

Value

The return value is a list of 6 components:

N.1 (or BA.1 or AGB.1) an array of abundance (or basal area or agb) in the first census; one dimension of the array for split1, the second for split2

N.2 (or BA.2 or AGB.2) abundance (or basal area or agb) in the second census in a matching array

date1 mean date of first census in a matching array

date2 mean date of second census in a matching array

interval the time interval in years in a matching array

little.r the rate of population change in a matching array, (log(N2) - log(N1))/time

This list can be submitted to assemble.demography() (topic utilitiesCTFS) to convert into a convenient table.

Details

Do not use this function with diameter categories as a split variable! The results won't make sense. The categories need to be permanent attributes, such as species, genus, quadrat. To find population change of dbh categories use instead pop.change.dbh()

Mean census date for a species is not the mean census date for all living individuals in that census, but the mean census date for all individuals alive in either census. Plants recruited between the two censuses get a first census date equal to the date on which the quadrat they later appear in was censused in the first census. Plants dead in the second census get a census date equal to the date on which their quadrat was censused

See also

abundance(), assemble.demography()

Examples

not_run({
  bcichange = pop.change(
    bci::bci12full5,
    bci::bci12full6,
    type = 'abund',
    split1 = bci::bci12full5$sp,
    mindbh = 10
  )
  str(bcichange)
  head(bcichange$N.1)
  change.table = assemble.demography(bcichange, type = 'a')
  head(change.table)
})