Change between two censuses in abundance or basal area.

Finds abundance or basal area in two censuses and the rate of change between them, in several dbh categories.

pop.change.dbh(census1, census2, type = "abund", dbhunit = "mm",
  alivecode = c("A"), classbreak = c(10, 100, 300), split = NULL)

Arguments

census1
The R Analytical Table for a single census, either tree or stem.

census2
The matching R Analytical Table for a later census.

type
Either 'abund' (default) for abundance, 'ba' for basal area, or 'agb'.

dbhunit
'cm' or 'mm', only used for basal area

alivecode
character, codes of the variable status that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems

classbreak
xxxdocparam

split
A vector of categorical values of the same length as census which groups trees into classes of interest for which abundance values are computed. This vector can be composed of characters or numbers.

Value

The return value is a list of two elements, one name abund (or ba) and the other meandate, just as other abundance results. Each is a table having one pair of columns for every dbh category: the first for census 1, the second for census 2. So if there are 3 dbh categories, the table has 6 columns. The rows of the table are the splitting variable (eg, species).

Details

Accepts two dataframes, each an R Analytical Table for one census, the earlier census first. Only one additional splitting variable (other than dbh category) is allowed. Typically, this is species, but genus or quadrat are other examples.

See also

Examples

# NOT RUN {
Nchange = pop.change.dbh(bciex::bci12t5mini, bciex::bci12t6mini, classbreak = c(10, 100, 300))
Nchange$abund

BAchangePerSpp <- pop.change.dbh(
  bciex::bci12t5mini,
  bciex::bci12t6mini,
  classbreak = c(10, 100),
  split = bciex::bci12t5mini$sp
)
head(BAchangePerSpp$ba)
# }