selectrandomquad

Draws rectangular quadrats in a plot at random. Returns a dataframe of coordinates defining the corners of all the quadrats.

The variable size is a vector of dimensions; shape allows a rectangle to be drawn; rep is the number of replicated random quadrats per dimension submitted.

Quadrats are chosen simply: a point is chosen by randomly drawing one x and one y coordinates to serve as the lower-left corner.

The x is chosen from a uniform distribution between 0 and plotdim1-size; y is chosen similarly.

All 3 algorithms for creating random quadrats under sample the plot corners. See selectrandomquad2() and selectrandomquad3(), alternative intended to overcome the bias (but neither does).

selectrandomquad(size, shape, rep, plotdim = c(1000, 500), graphit = FALSE)

Arguments

size

a vector of quadrat sizes, referring to the x-dimension of a rectangular quadrat

shape

the ratio of y to x dimensions of the rectangles; rectdim=1 (the default) for squares

rep

the number of random quadrats to create, of each size

plotdim

the x and y dimensions of the entire plot

graphit

whether to graph the locations of the chosen quadrats on a plot map

censdata

one R Analytical Table, either full or stem