trim.growth

This is where growth rates are excluded. It is based on a linear model estimating the standard deviation of dbh measures (due to error, that is); the parameters slope and intercept define the linear relationship between this error deviation and dbh. Any case where the second dbh measure is more than 4 standard deviations below the first is marked false, meaning it will be excluded from growth measurements. The default values of slope and intercept are based on dbh remeasure tests done in both 1995 and 2000 at BCI. A line was fitted through the absolute dbh errors as a function of dbh in both years; the average slope and intercept is used here. The function also excludes any growth rate > 75 mm per yr, cases where the stemID changes, and if the POM changes by more than 5 All parameters for excluding growth measures based on error can be adjusted: to include all measures, set maxgrow and err.limit to very high numbers, such as 10000; to include POM changes, set pomcut to a high number, such as 10; to include cases where stemID changed, set exclude.stem.change=FALSE. This function is usually only used inside the other growth functions. With exclude.stem.change==FALSE, keep all cases where stem changes, regardless of growth (it does not make sense to exclude a record based on growth when the growth is based on different stems). Note that trees are exclude if cens1$dbh<mindbh, but not if cens2$dbh<mindbh.

trim.growth(cens1, cens2, time, slope = 0.006214, intercept = 0.9036,
  err.limit = 4, maxgrow = 75, pomcut = 0.05, mindbh = 10,
  dbhunit = "mm", exclude.stem.change = TRUE)