wavelet.allsp

Function to calculate the wavelet variance curves for all species in one plot using quadrat indexation from CTFS package.

wavelet.allsp(censdata, plotdim = c(1000, 500), gridsize = 2.5,
  mindbh = NULL)

Examples

not_run({
  load("bci.full1.rdata")
  wavelet.variances = wavelet.allsp(censdata, plotdim=c(1000,500))
})