Create a dummy wsgdata data set

Create a dummy wsgdata data set with random values of species density.

wsgdata_dummy(censdata, plot = "bci")

Arguments

censdata
Census data, e.g. bci::bci12full1.

plot
Character string giving the name of the plot where the data comes from (as it appears in the names of the R Analytical Tables).

Value

A data frame with one row per species in censdata and 3 variables.

Details

If you lack a real wood density table, you can use this function to create a dummy one. For example, this may help you to test biomass.CTFSdb() and density_ind().

See ?density_ind(), particularly notice the following: "There must be a table of wood density submitted (wsgdata), and this table must have a column sp with species names, a column plot, plus the wood density in a column called wsg (though the name of that column can be changed using the argument denscol)".

Examples

wsgdata_dummy(bci::bci12stem1)

#>         sp plot           wsg
#> 1   swars1  bci  2.419707e-01
#> 2   hybapr  bci  5.399097e-02
#> 3   aegipa  bci  4.431848e-03
#> 4   beilpe  bci  1.338302e-04
#> 5   faraoc  bci  1.486720e-06
#> 6   tet2pa  bci  6.075883e-09
#> 7   des2pa  bci  9.134720e-12
#> 8   crotbi  bci  5.052271e-15
#> 9   alsebl  bci  1.027977e-18
#> 10  ouralu  bci  7.694599e-23
#> 11  mourmy  bci  2.118819e-27
#> 12  ocotob  bci  2.146384e-32
#> 13  paligu  bci  7.998828e-38
#> 14  solaha  bci  1.096607e-43
#> 15  quaras  bci  5.530710e-50
#> 16  gustsu  bci  1.026163e-56
#> 17  pterro  bci  7.004182e-64
#> 18  virosu  bci  1.758750e-71
#> 19  simaam  bci  1.624636e-79
#> 20  protte  bci  5.520948e-88
#> 21  brosal  bci  6.902029e-97
#> 22  ingape  bci 3.174282e-106
#> 23  poular  bci 5.370560e-116
#> 24  casear  bci 3.342714e-126
#> 25  hirttr  bci 7.653930e-137
#> 26  ocotwh  bci 6.447260e-148
#> 27  guatdu  bci 1.997889e-159
#> 28  anacex  bci 2.277577e-171
#> 29  tri2tu  bci 9.551695e-184
#> 30  guarsp  bci 1.473646e-196
#> 31  guargu  bci 8.363952e-210
#> 32  zantpr  bci 1.746366e-223
#> 33  termob  bci 1.341420e-237
#> 34  guazul  bci 3.790526e-252
#> 35  tab1ro  bci 3.940396e-267
#> 36  sloate  bci 1.506905e-282
#> 37  alchco  bci 2.120007e-298
#> 38  loncla  bci 1.097221e-314
#> 39  luehse  bci  0.000000e+00
#> 40  tripcu  bci  0.000000e+00
#> 41  cecrob  bci  0.000000e+00
#> 42  licapl  bci  0.000000e+00
#> 43  eugeco  bci  0.000000e+00
#> 44  heisco  bci  0.000000e+00
#> 45  eugene  bci  0.000000e+00
#> 46  pri2co  bci  0.000000e+00
#> 47  sponra  bci  0.000000e+00
#> 48  poutre  bci  0.000000e+00
#> 49  guapst  bci  0.000000e+00
#> 50  cordbi  bci  0.000000e+00
#> 51  zantbe  bci  0.000000e+00
#> 52  annosp  bci  0.000000e+00
#> 53  tab2ar  bci  0.000000e+00
#> 54  jac1co  bci  0.000000e+00
#> 55  zantp1  bci  0.000000e+00
#> 56  chr2ar  bci  0.000000e+00
#> 57  eugeoe  bci  0.000000e+00
#> 58  huracr  bci  0.000000e+00
#> 59  sapiau  bci  0.000000e+00
#> 60  apeime  bci  0.000000e+00
#> 61  ficuyo  bci  0.000000e+00
#> 62  ingago  bci  0.000000e+00
#> 63  ceibpe  bci  0.000000e+00
#> 64  cavapl  bci  0.000000e+00
#> 65  chr2ca  bci  0.000000e+00
#> 66  lindla  bci  0.000000e+00
#> 67  virose  bci  0.000000e+00
#> 68  pla1pi  bci  0.000000e+00
#> 69  lacmpa  bci  0.000000e+00
#> 70  drypst  bci  0.000000e+00
#> 71  myrofr  bci  0.000000e+00
#> 72  ingaco  bci  0.000000e+00
#> 73  posola  bci  0.000000e+00
#> 74  casesy  bci  0.000000e+00
#> 75  amaico  bci  0.000000e+00
#> 76  pla2el  bci  0.000000e+00
#> 77  cordal  bci  0.000000e+00
#> 78  dio2ar  bci  0.000000e+00
#> 79  tratas  bci  0.000000e+00
#> 80  colugl  bci  0.000000e+00
#> 81  apeiti  bci  0.000000e+00
#> 82  gar2in  bci  0.000000e+00
#> 83  poutst  bci  0.000000e+00
#> 84  tab1gu  bci  0.000000e+00
#> 85  ingafa  bci  0.000000e+00
#> 86  sch1zo  bci  0.000000e+00
#> 87  sterap  bci  0.000000e+00
#> 88  cassel  bci  0.000000e+00
#> 89  calolo  bci  0.000000e+00
#> 90  hyeral  bci  0.000000e+00
#> 91  zuelgu  bci  0.000000e+00
#> 92  cecrin  bci  0.000000e+00
#> 93  hassfl  bci  0.000000e+00
#> 94  sponmo  bci  0.000000e+00
#> 95  ingama  bci  0.000000e+00
#> 96  allops  bci  0.000000e+00
#> 97  virosp  bci  0.000000e+00
#> 98  sympgl  bci  0.000000e+00
#> 99  guetfo  bci  0.000000e+00
#> 100 mar1la  bci  0.000000e+00
#> 101 ficutr  bci  0.000000e+00
#> 102 aspicr  bci  0.000000e+00
#> 103 dendar  bci  0.000000e+00
#> 104 theoca  bci  0.000000e+00
#> 105 andiin  bci  0.000000e+00
#> 106 ingas1  bci  0.000000e+00
#> 107 diptpa  bci  0.000000e+00
#> 108 ast2gr  bci  0.000000e+00
#> 109 hirtam  bci  0.000000e+00
#> 110 ficuob  bci  0.000000e+00
#> 111 tropra  bci  0.000000e+00
#> 112 protpa  bci  0.000000e+00
#> 113 ficuto  bci  0.000000e+00
#> 114 unonpi  bci  0.000000e+00
#> 115 maquco  bci  0.000000e+00
#> 116 gar2ma  bci  0.000000e+00
#> 117 macrgl  bci  0.000000e+00
#> 118 randar  bci  0.000000e+00
#> 119 cespma  bci  0.000000e+00
#> 120 cordla  bci  0.000000e+00
#> 121 picrla  bci  0.000000e+00
#> 122 cupala  bci  0.000000e+00
#> 123 zantse  bci  0.000000e+00
#> 124 xyl1ma  bci  0.000000e+00
#> 125 geniam  bci  0.000000e+00
#> 126 ingasp  bci  0.000000e+00
#> 127 hampap  bci  0.000000e+00
#> 128 sapisp  bci  0.000000e+00
#> 129 sipapa  bci  0.000000e+00
#> 130 myrcga  bci  0.000000e+00
#> 131 nectci  bci  0.000000e+00
#> 132 ficuma  bci  0.000000e+00
#> 133 maytsc  bci  0.000000e+00
#> 134 phoeci  bci  0.000000e+00
#> 135 licahy  bci  0.000000e+00
#> 136 tachve  bci  0.000000e+00
#> 137 protco  bci  0.000000e+00
#> 138 laciag  bci  0.000000e+00
#> 139 psidfr  bci  0.000000e+00
#> 140 celtsc  bci  0.000000e+00
#> 141 nects1  bci  0.000000e+00
#> 142 ormocr  bci  0.000000e+00
#> 143 apeihy  bci  0.000000e+00
#> 144 ficupo  bci  0.000000e+00
#> 145 turpoc  bci  0.000000e+00
#> 146 ery1co  bci  0.000000e+00
#> 147 tremmi  bci  0.000000e+00
#> 148 ade1tr  bci  0.000000e+00
#> 149 ingasa  bci  0.000000e+00
#> 150 ocotce  bci  0.000000e+00
#> 151 guargr  bci  0.000000e+00
#> 152 ingapu  bci  0.000000e+00
#> 153 nectgl  bci  0.000000e+00
#> 154 swars2  bci  0.000000e+00
#> 155 micoar  bci  0.000000e+00
#> 156 ficuc2  bci  0.000000e+00
#> 157 ficuin  bci  0.000000e+00
#> 158 ormoma  bci  0.000000e+00
#> 159 termam  bci  0.000000e+00
#> 160 tri2pa  bci  0.000000e+00
#> 161 ficubu  bci  0.000000e+00
#> 162 tet4jo  bci  0.000000e+00
#> 163 pse1se  bci  0.000000e+00
#> 164 tocopi  bci  0.000000e+00
#> 165 sch2mo  bci  0.000000e+00
#> 166 pochse  bci  0.000000e+00
#> 167 ocotpu  bci  0.000000e+00
#> 168 cupasy  bci  0.000000e+00
#> 169 ingami  bci  0.000000e+00
#> 170 ingaru  bci  0.000000e+00
#> 171 cuparu  bci  0.000000e+00
#> 172 poutfo  bci  0.000000e+00
#> 173 vochfe  bci  0.000000e+00
#> 174 ast1st  bci  0.000000e+00
#> 175 cou2cu  bci  0.000000e+00
#> 176 micoho  bci  0.000000e+00
#> 177 coccco  bci  0.000000e+00
#> 178 spacme  bci  0.000000e+00
#> 179 micoel  bci  0.000000e+00
#> 180 cupaci  bci  0.000000e+00
#> 181 eugega  bci  0.000000e+00
#> 182 tri4ga  bci  0.000000e+00
#> 183 laetpr  bci  0.000000e+00
#> 184 lafopu  bci  0.000000e+00
#> 185 entesc  bci  0.000000e+00
#> 186 pterof  bci  0.000000e+00
#> 187 cedrod  bci  0.000000e+00
#> 188 ingaqu  bci  0.000000e+00
#> 189 pourbi  bci  0.000000e+00
#> 190 pit1ma  bci  0.000000e+00
#> 191 uniden  bci  0.000000e+00
#> 192 ingath  bci  0.000000e+00
#> 193 pochqu  bci  0.000000e+00
#> 194 alchla  bci  0.000000e+00
#> 195 sipagu  bci  0.000000e+00
#> 196 ochrpy  bci  0.000000e+00
#> 197 psycde  bci  0.000000e+00
#> 198 ardife  bci  0.000000e+00
#> 199 psycho  bci  0.000000e+00
#> 200 acaldi  bci  0.000000e+00
#> 201 pentma  bci  0.000000e+00
#> 202 annoac  bci  0.000000e+00
#> 203 oenoma  bci  0.000000e+00
#> 204 psycma  bci  0.000000e+00
#> 205 taline  bci  0.000000e+00
#> 206 soroaf  bci  0.000000e+00
#> 207 talipr  bci  0.000000e+00
#> 208 malmsp  bci  0.000000e+00
#> 209 caseac  bci  0.000000e+00
#> 210 olmeas  bci  0.000000e+00
#> 211 ingaum  bci  0.000000e+00
#> 212 rinosy  bci  0.000000e+00
#> 213 cappfr  bci  0.000000e+00
#> 214 stylst  bci  0.000000e+00
#> 215 ery2pa  bci  0.000000e+00
#> 216 herrpu  bci  0.000000e+00
#> 217 laetth  bci  0.000000e+00
#> 218 ery2ma  bci  0.000000e+00
#> 219 bactc1  bci  0.000000e+00
#> 220 alibed  bci  0.000000e+00
#> 221 psycg3  bci  0.000000e+00
#> 222 chr1ec  bci  0.000000e+00
#> 223 cestme  bci  0.000000e+00
#> 224 coccma  bci  0.000000e+00
#> 225 bactba  bci  0.000000e+00
#> 226 heisac  bci  0.000000e+00
#> 227 pipeco  bci  0.000000e+00
#> 228 vismba  bci  0.000000e+00
#> 229 sennda  bci  0.000000e+00
#> 230 pipeae  bci  0.000000e+00
#> 231 quasam  bci  0.000000e+00
#> 232 conoci  bci  0.000000e+00
#> 233 nectpu  bci  0.000000e+00
#> 234 pipea1  bci  0.000000e+00
#> 235 socrex  bci  0.000000e+00
#> 236 aphesi  bci  0.000000e+00
#> 237 pipecu  bci  0.000000e+00
#> 238 thevah  bci  0.000000e+00
#> 239 perexa  bci  0.000000e+00
#> 240 geonin  bci  0.000000e+00
#> 241 hameax  bci  0.000000e+00
#> 242 pipere  bci  0.000000e+00
#> 243 psycpi  bci  0.000000e+00
#> 244 clidoc  bci  0.000000e+00
#> 245 bactma  bci  0.000000e+00
#> 246 malpro  bci  0.000000e+00
#> 247 pipeca  bci  0.000000e+00
#> 248 micoaf  bci  0.000000e+00
#> 249 pipepe  bci  0.000000e+00
#> 250 neeaam  bci  0.000000e+00
#> 251 randfo  bci  0.000000e+00
#> 252 xyl2ol  bci  0.000000e+00
#> 253 ardiba  bci  0.000000e+00
#> 254 brosgu  bci  0.000000e+00
#> 255 margno  bci  0.000000e+00
#> 256 casegu  bci  0.000000e+00
#> 257 bactc2  bci  0.000000e+00
#> 258 psycli  bci  0.000000e+00
#> 259 solast  bci  0.000000e+00
#> 260 psycch  bci  0.000000e+00
#> 261 chimpa  bci  0.000000e+00
#> 262 lozapi  bci  0.000000e+00
#> 263 conobr  bci  0.000000e+00
#> 264 bertgu  bci  0.000000e+00
#> 265 pit1ru  bci  0.000000e+00
#> 266 cnempe  bci  0.000000e+00
#> 267 micoim  bci  0.000000e+00
#> 268 cha1te  bci  0.000000e+00
#> 269 micone  bci  0.000000e+00
#> 270 lycima  bci  0.000000e+00
#> 271 pipeim  bci  0.000000e+00
#> 272 protsp  bci  0.000000e+00
#> 273 couthe  bci  0.000000e+00
#> 274 appuse  bci  0.000000e+00
#> 275 clidde  bci  0.000000e+00
#> 276 acalma  bci  0.000000e+00
#> 277 urerba  bci  0.000000e+00
#> 278 psycg2  bci  0.000000e+00
#> 279 vismbi  bci  0.000000e+00
#> 280 stemgr  bci  0.000000e+00
#> 281 acacme  bci  0.000000e+00
#> 282 ormoam  bci  0.000000e+00
#> 283 tri1gi  bci  0.000000e+00
#> 284 psycfu  bci  0.000000e+00
#> 285 cha2sc  bci  0.000000e+00
#> 286 ternte  bci  0.000000e+00
#> 287 annoha  bci  0.000000e+00
#> 288 psycac  bci  0.000000e+00
#> 289 schipa  bci  0.000000e+00
#> 290 elaeol  bci  0.000000e+00
#> 291 ardigu  bci  0.000000e+00
#> 292 vismma  bci  0.000000e+00
#> 293 maclti  bci  0.000000e+00
#> 294 psycg1  bci  0.000000e+00
#> 295 cyphha  bci  0.000000e+00
#> 296 boropa  bci  0.000000e+00
#> 297 nects3  bci  0.000000e+00
#> 298 caseco  bci  0.000000e+00
#> 299 koanwe  bci  0.000000e+00
#> 300 clidse  bci  0.000000e+00
#> 301 psycra  bci  0.000000e+00
#> 302 lopida  bci  0.000000e+00
#> 303 ingam1  bci  0.000000e+00
#> 304 ficuc1  bci  0.000000e+00
#> 305 anaxpa  bci  0.000000e+00
#> 306 solaar  bci  0.000000e+00
#> 307 xyl2ch  bci  0.000000e+00
#> 308 hamepa  bci  0.000000e+00
#> 309 leandi  bci  0.000000e+00
#> 310 micopr  bci  0.000000e+00
#> 311 banagu  bci  0.000000e+00
#> 312 solaas  bci  0.000000e+00
#> 313 ficuci  bci  0.000000e+00
#> 314 psycb1  bci  0.000000e+00
#> 315 ficupe  bci  0.000000e+00
#> 316 tremin  bci  0.000000e+00
#> 317 psycps  bci  0.000000e+00
#> 318 cecrlo  bci  0.000000e+00
#> 319 rauvli  bci  0.000000e+00
#> 320 verbgi  bci  0.000000e+00
#> 321 micodo  bci  0.000000e+00
#> 322 carica  bci  0.000000e+00
#> 323 oreoca  bci  0.000000e+00