Create a dummy wsgdata data set

Create a dummy wsgdata data set with random values of species density.

wsgdata_dummy(censdata, plot = "bci")

Arguments

censdata
Census data, e.g. bci::bci12full1.

plot
Character string giving the name of the plot where the data comes from (as it appears in the names of the R Analytical Tables).

Value

A data frame with one row per species in censdata and 3 variables.

Details

If you lack a real wood density table, you can use this function to create a dummy one. For example, this may help you to test biomass.CTFSdb() and density_ind().

See ?density_ind(), particularly notice the following: "There must be a table of wood density submitted (wsgdata), and this table must have a column sp with species names, a column plot, plus the wood density in a column called wsg (though the name of that column can be changed using the argument denscol)".

Examples

wsgdata_dummy(bci::bci12stem1)

#> Error in loadNamespace(name): there is no package called ‘bci’